requirements.txt
absl-py==2.2.2
accelerate==1.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
archspec==0.2.5
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==3.3.10
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.1.0
audioread==3.0.1
babel==2.17.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
boltons==24.0.0
Brotli==1.1.0
cachetools==5.5.2
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
cmake==3.31.4
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.2
conda==25.1.0
conda-build==25.1.1
conda_index==0.5.0
conda-libmamba-solver==25.1.1
conda-package-handling==2.4.0
conda_package_streaming==0.11.0
contourpy==1.3.2
# cosine_annealing_warmup==2.0
cssselect2==0.8.0
cycler==0.12.1
databricks-sdk==0.54.0
DateTime==5.5
debugpy==1.8.14
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.4.0
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
einops==0.8.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
expecttest==0.3.0
fastjsonschema==2.21.1
filelock==3.17.0
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.0
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.2
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.2
grpcio==1.71.0
gunicorn==23.0.0
h11==0.16.0
h2==4.1.0
h5py==3.13.0
hf-xet==1.1.2
hpack==4.0.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.1
humanfriendly==10.0
hyperframe==6.0.1
hypothesis==6.124.7
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.7
isoduration==20.11.0
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.5.1
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter_client==8.6.3
jupyter-console==6.6.3
jupyter_core==5.8.0
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
lazy_loader==0.4
libarchive-c==5.1
libmambapy==2.0.5
librosa==0.10.2.post1
lief==0.14.1
lightgbm==4.6.0
lightning==2.5.0.post0
lightning-utilities==0.14.3
lintrunner==0.12.7
llvmlite==0.44.0
Mako==1.3.10
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mccabe==0.7.0
menuinst==2.2.0
mergedeep==1.3.4
mistune==3.1.3
mlflow==2.20.3
mlflow-skinny==2.20.3
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
notebook==7.4.3
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnx==1.17.0
onnxruntime==1.21.0
openpyxl==3.1.5
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
optree==0.14.0
optuna==4.2.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pip==24.3.1
pkginfo==1.12.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
prometheus_client==0.22.0
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.4
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycosat==0.6.6
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pylint==3.3.5
pymongo==4.11.2
pyparsing==3.2.3
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-etcd==0.4.5
python-json-logger==3.3.0
pytorch-lightning==2.5.1.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.4.0
ray==2.43.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
rsa==4.9.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.8
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.5
soxr==0.5.0.post1
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack_data==0.6.3
subprocess.run==0.0.8
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.5.26
tinycss2==1.4.0
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchelastic==0.2.2
torchmetrics==1.7.1
torchsummary==1.5.1
torchvision==0.21.0+cu126
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
triton==3.2.0
truststore==0.10.0
types-dataclasses==0.6.6
types-python-dateutil==2.9.0.20250516
typing_extensions==4.12.2
typing-inspection==0.4.1
tzdata==2025.2
uri-template==1.3.0
urllib3==2.3.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
wrapt==1.17.2
xmltodict==0.14.2
yacs==0.1.8
yarl==1.20.0
zipp==3.21.0
zope.interface==7.2
zstandard==0.23.0
